Main parts of the game finished
Players are now able to play the game as many times as they want.
Splash scene and menu scene are now different scenes.
Game Over scene is included in the Game Scene. It shows the player
score, a random phrase and give the player the possibility to play again
or to leave to the menu.